requirements.txt
qrcode
opencv-python
pyzbar
Pillow
argparse